refactor: split UsageViewModel into focused extensions (761 → 457 lines)
UsageViewModel.swift had grown to 761 lines, just shy of the 800-line cap
this codebase enforces. Three cohesive method clusters peeled off cleanly
into same-type extensions without changing behavior:

- UsageViewModel+Statics.swift (94 lines) — the existing
  `MARK: Testable static helpers` block. `clampedRefreshInterval`,
  `refreshErrorMessage`, `hasDataChanged`, `effectiveRateLimits`,
  `effectiveValue`, and the three nonisolated refresh-interval constants.
  All `nonisolated static`; no behavior or signature change. Throttle
  bookkeeping (`recordThrottleEvent`/`throttleCount`) stays in the main
  file because it owns a stored `private static var throttleTracker`.

- UsageViewModel+Lifecycle.swift (185 lines) — file-watcher setup,
  sleep/wake/screen-lock observers, idle-suspend / activity-monitor
  resume plumbing, and the polling timer (`startPolling`,
  `restartPolling`, `updatePollingInterval`, `refreshInterval`). None
  of these touch `@Published` state directly — they only schedule work
  and call back into `refresh()`. `deinit` cleanup stays in the main
  file (Swift doesn't allow `deinit` in extensions).

- UsageViewModel+FanOut.swift (78 lines) — `scheduleFanOut` and
  `fetchAllAccounts`. Triggered by the "Show all accounts in menu bar"
  preference observer in `init` (debounced, then `scheduleFanOut()`).
  The single in-flight task is tracked via `pendingFanOut` on the main
  type so back-to-back triggers coalesce.

Cross-file visibility:

Swift `private` is file-scoped; symbols touched by extension files
were widened to internal (module-private). Each widened symbol is
documented inline with the extension that reads/writes it:

- Stored state read/written across files: `aggregator`, `fileWatcher`,
  `pollingTimer`, `wakeObserver`, `sleepObserver`, `lockObserver`,
  `unlockObserver`, `adaptivePolling`, `isSuspended`, `pendingFanOut`,
  and `aggregateOffMain(...)`.
- `perAccountRateLimits` lost `private(set)` so the fan-out extension
  can write to it; nothing outside the module touches it.

After:
- UsageViewModel.swift:           457 lines
- UsageViewModel+Lifecycle.swift: 185 lines
- UsageViewModel+Statics.swift:    94 lines
- UsageViewModel+FanOut.swift:     78 lines
- (combined): 814 lines (was 761, +53 from per-file imports + section comments)

No production behavior change. Test suite unchanged.
